bug: Fix mutual exclusivity between name and file
This resolves the state where a user has supplied the `--file`/`-f`
and also the `--name`/`-n` command line options.

Previously if `--file` was supplied the code path would use the
`batch_create()` function which was hard coded to use the filename
`voice.mp3`.  This re-works the definition of `batch_create()` to both
use an annotated argument as well as updating calls to supply the
filename argument.